rename avatar -> UserAvatar
