+++ resources/js/components/header/TopResultCard.vue
@@ -4,7 +4,7 @@ import BuildingOfficeIcon from '../ui/icons/micro/BuildingOfficeIcon.vue'
 import UserIcon from '../ui/icons/micro/UserIcon.vue'
 import BriefcaseIcon from '../ui/icons/micro/BriefcaseIcon.vue'
 import OrganizationLogo from '../ui/OrganizationLogo.vue'
-import Avatar from '../ui/Avatar.vue'
+import Avatar from '../ui/UserAvatar.vue'
 import UserGroupIcon from '../ui/icons/mini/UserGroupIcon.vue'
 import { formatDealAmount, getDealStatusClass } from '../../utils'
